chore: drop viz wrappers
